Fixed a bug when the testing would end before finishing all tasks
Also changed the build script to avoid GCC optimizations (except for the system architecture). The optimization is almost unbelievable, probably because the code isn't doing anything useful and is highly predictable but doing so defeats the purpose of the test. As this is not a tool meant to be used as a real benchmarking tool, will not hurt that much. But if you still want to see your PC get 2k+ ST points go ahead and build with -O3 flag xD